Add latency test with client/server
