+++ public/index.html
@@ -39,5 +39,20 @@
       To begin the development, run `npm start` or `yarn start`.
       To create a production bundle, use `npm run build` or `yarn build`.
     -->
+    <script src="/socket.io/socket.io.js"></script>
+    <script>
+      var socket = io();
+
+      setInterval(() => {
+        const start = Date.now();
+
+        // volatile, so the packet will be discarded if the socket is not connected
+        socket.volatile.emit("ping", () => {
+          const latency = Date.now() - start;
+          console.log(latency + " ms")
+          // ...
+        });
+      }, 5000);
+    </script>
   </body>
 </html>
